# Add microchip lan865x macphy fullduplex mode support