--- a/drivers/ethernet/Kconfig.oa_tc6
+++ b/drivers/ethernet/Kconfig.oa_tc6
@@ -34,10 +34,15 @@ config OA_TC6_TIMEOUT
 config OA_TC6_TX_RX_BUFFER_SIZE
 	int "OA SPI TX RX BUFFER SIZE"
 	default 31
+	range 1 31
 	help
 	  Size of the buffer to transmit and receive SPI chunks to MAC-PHY
 	  and from MAC-PHY respectively. One buffer size is equal to 68
 	  bytes. (Size of the single OASPI chunk is 68 bytes). Buffer size
-	  31 is equal to (31 * 68) 2108 bytes.
+	  31 is equal to (31 * 68) 2108 bytes. In full duplex mode same
+	  buffer size should be used in TX and RX path. For SoCs with more
+	  resources, set the maximum value of 31 (2108 bytes) to achieve
+	  higher bandwidth. For SoCs with limited resources, set the value
+	  to at least 1 (68 bytes).
 
 endif