--- a/drivers/ethernet/eth_lan865x.c
+++ b/drivers/ethernet/eth_lan865x.c
@@ -64,29 +64,11 @@ static int lan865x_mac_rxtx_control(const struct device *dev, bool en)
 	return oa_tc6_reg_write(ctx->tc6, LAN865x_MAC_NCR, ctl);
 }
 
-static int lan865x_enable_sync(const struct device *dev)
-{
-	struct lan865x_data *ctx = dev->data;
-	uint32_t val;
-	int ret;
-
-	ret = oa_tc6_reg_read(ctx->tc6, OA_CONFIG0, &val);
-	if (ret) {
-		return ret;
-	}
-	val |= OA_CONFIG0_SYNC | OA_CONFIG0_RFA_ZARFE;
-	ret = oa_tc6_reg_write(ctx->tc6, OA_CONFIG0, val);
-	if (ret) {
-		return ret;
-	}
-
-	return lan865x_mac_rxtx_control(dev, LAN865x_MAC_TXRX_ON);
-}
-
 static void lan865x_iface_init(struct net_if *iface)
 {
 	const struct device *dev = net_if_get_device(iface);
 	struct lan865x_data *ctx = dev->data;
+	struct oa_tc6 *tc6 = ctx->tc6;
 	int ret;
 
 #if defined(CONFIG_NET_VLAN)
@@ -99,20 +81,31 @@ static void lan865x_iface_init(struct net_if *iface)
 	}
 #endif
 
-	ret = lan865x_enable_sync(dev);
+	ret = oa_tc6_enable_sync(tc6);
 	if (ret) {
-		LOG_ERR("LAN865x sync enable failed: %d\n", ret);
+		LOG_ERR("Sync enable failed: %d\n", ret);
+		return;
+	}
+
+	ret = oa_tc6_zero_align_receive_frame_enable(tc6);
+	if (ret) {
+		LOG_ERR("ZARFE enable failed: %d\n", ret);
+		return;
+	}
+
+	ret = lan865x_mac_rxtx_control(dev, LAN865x_MAC_TXRX_ON);
+	if (ret) {
+		LOG_ERR("LAN865x TX and RX enable failed: %d\n", ret);
 		return;
 	}
 
 	net_if_set_link_addr(iface, ctx->mac_address, sizeof(ctx->mac_address), NET_LINK_ETHERNET);
 
-	ctx->iface = iface;
+	tc6->iface = iface;
 
 	ethernet_init(iface);
 
 	net_eth_carrier_on(iface);
-	ctx->iface_initialized = true;
 }
 
 static enum ethernet_hw_caps lan865x_port_get_capabilities(const struct device *dev __unused,
@@ -125,7 +118,6 @@ static enum ethernet_hw_caps lan865x_port_get_capabilities(const struct device *
 		;
 }
 
-static int lan865x_gpio_reset(const struct device *dev);
 static void lan865x_write_macaddress(const struct device *dev);
 static int lan865x_set_config(const struct device *dev,
 			      struct net_if *iface __unused,
@@ -147,42 +139,6 @@ static int lan865x_set_config(const struct device *dev,
 	}
 
 	return -ENOTSUP;
-}
-
-static int lan865x_wait_for_reset(const struct device *dev)
-{
-	struct lan865x_data *ctx = dev->data;
-	uint8_t i;
-
-	/* Wait for end of LAN865x reset */
-	for (i = 0; !ctx->reset && i < LAN865X_RESET_TIMEOUT; i++) {
-		k_msleep(1);
-	}
-
-	if (i == LAN865X_RESET_TIMEOUT) {
-		LOG_ERR("LAN865x reset timeout reached!");
-		return -ENODEV;
-	}
-
-	return 0;
-}
-
-static int lan865x_gpio_reset(const struct device *dev)
-{
-	const struct lan865x_config *cfg = dev->config;
-	struct lan865x_data *ctx = dev->data;
-
-	ctx->reset = false;
-	ctx->tc6->protected = false;
-
-	/* Perform (GPIO based) HW reset */
-	/* assert RESET_N low for 10 µs (5 µs min) */
-	gpio_pin_set_dt(&cfg->reset, 1);
-	k_busy_wait(10U);
-	/* deassert - end of reset indicated by IRQ_N low  */
-	gpio_pin_set_dt(&cfg->reset, 0);
-
-	return lan865x_wait_for_reset(dev);
 }
 
 static int lan865x_check_spi(const struct device *dev)
@@ -274,153 +230,23 @@ static int lan865x_default_config(const struct device *dev)
 	return 0;
 }
 
-static void lan865x_int_callback(const struct device *dev, struct gpio_callback *cb, uint32_t pins)
-{
-	ARG_UNUSED(dev);
-	ARG_UNUSED(pins);
-
-	struct lan865x_data *ctx = CONTAINER_OF(cb, struct lan865x_data, gpio_int_callback);
-
-	k_sem_give(&ctx->int_sem);
-}
-
-static void lan865x_read_chunks(const struct device *dev)
-{
-	struct lan865x_data *ctx = dev->data;
-	struct oa_tc6 *tc6 = ctx->tc6;
-	struct net_pkt *pkt;
-	int ret;
-
-	pkt = net_pkt_rx_alloc(K_MSEC(CONFIG_ETH_LAN865X_TIMEOUT));
-	if (!pkt) {
-		LOG_ERR("OA RX: Could not allocate packet!");
-		return;
-	}
-
-	k_sem_take(&ctx->tx_rx_sem, K_FOREVER);
-	ret = oa_tc6_read_chunks(tc6, pkt);
-	if (ret < 0) {
-		eth_stats_update_errors_rx(ctx->iface);
-		net_pkt_unref(pkt);
-		k_sem_give(&ctx->tx_rx_sem);
-		return;
-	}
-
-	/* Feed buffer frame to IP stack */
-	ret = net_recv_data(ctx->iface, pkt);
-	if (ret < 0) {
-		LOG_ERR("OA RX: Could not process packet (%d)!", ret);
-		net_pkt_unref(pkt);
-	}
-	k_sem_give(&ctx->tx_rx_sem);
-}
-
-static void lan865x_int_thread(const struct device *dev)
-{
-	struct lan865x_data *ctx = dev->data;
-	struct oa_tc6 *tc6 = ctx->tc6;
-	uint32_t sts, ftr;
-	int ret;
-
-	while (true) {
-		k_sem_take(&ctx->int_sem, K_FOREVER);
-		if (!ctx->reset) {
-			oa_tc6_reg_read(tc6, OA_STATUS0, &sts);
-			if (sts & OA_STATUS0_RESETC) {
-				oa_tc6_reg_write(tc6, OA_STATUS0, sts);
-
-				lan865x_default_config(dev);
-
-				ctx->reset = true;
-				/*
-				 * According to OA T1S standard - it is mandatory to
-				 * read chunk of data to get the IRQ_N negated (deasserted).
-				 */
-				oa_tc6_read_status(tc6, &ftr);
-				continue;
-			}
-		}
-
-		/*
-		 * The IRQ_N is asserted when RCA becomes > 0. As described in
-		 * OPEN Alliance 10BASE-T1x standard it is deasserted when first
-		 * data header is received by LAN865x.
-		 *
-		 * Hence, it is mandatory to ALWAYS read at least one data chunk!
-		 */
-		do {
-			lan865x_read_chunks(dev);
-		} while (tc6->rca > 0);
-
-		ret = oa_tc6_check_status(tc6);
-		if (ret == -EIO) {
-			lan865x_gpio_reset(dev);
-		}
-	}
-}
-
 static int lan865x_init(const struct device *dev)
 {
 	const struct lan865x_config *cfg = dev->config;
 	struct lan865x_data *ctx = dev->data;
+	struct oa_tc6 *tc6 = ctx->tc6;
 	int ret;
 
-	__ASSERT(cfg->spi.config.frequency <= LAN865X_SPI_MAX_FREQUENCY,
-		 "SPI frequency exceeds supported maximum\n");
-
-	if (!spi_is_ready_dt(&cfg->spi)) {
-		LOG_ERR("SPI bus %s not ready", cfg->spi.bus->name);
-		return -ENODEV;
-	}
-
-	if (!gpio_is_ready_dt(&cfg->interrupt)) {
-		LOG_ERR("Interrupt GPIO device %s is not ready", cfg->interrupt.port->name);
-		return -ENODEV;
+	ret = oa_tc6_init(tc6);
+	if (ret < 0) {
+		LOG_ERR("OA TC6 init failed %d", ret);
+		return ret;
 	}
 
 	/* Check SPI communication after reset */
 	ret = lan865x_check_spi(dev);
 	if (ret < 0) {
-		LOG_ERR("SPI communication not working, %d", ret);
-		return ret;
-	}
-
-	/*
-	 * Configure interrupt service routine for LAN865x IRQ
-	 */
-	ret = gpio_pin_configure_dt(&cfg->interrupt, GPIO_INPUT);
-	if (ret < 0) {
-		LOG_ERR("Failed to configure interrupt GPIO, %d", ret);
-		return ret;
-	}
-
-	gpio_init_callback(&(ctx->gpio_int_callback), lan865x_int_callback,
-			   BIT(cfg->interrupt.pin));
-
-	ret = gpio_add_callback(cfg->interrupt.port, &ctx->gpio_int_callback);
-	if (ret < 0) {
-		LOG_ERR("Failed to add INT callback, %d", ret);
-		return ret;
-	}
-
-	gpio_pin_interrupt_configure_dt(&cfg->interrupt, GPIO_INT_EDGE_TO_ACTIVE);
-
-	/* Start interruption-poll thread */
-	ctx->tid_int = k_thread_create(
-		&ctx->thread, ctx->thread_stack, CONFIG_ETH_LAN865X_IRQ_THREAD_STACK_SIZE,
-		(k_thread_entry_t)lan865x_int_thread, (void *)dev, NULL, NULL,
-		K_PRIO_COOP(CONFIG_ETH_LAN865X_IRQ_THREAD_PRIO), 0, K_NO_WAIT);
-	k_thread_name_set(ctx->tid_int, "lan865x_interrupt");
-
-	/* Perform HW reset - 'rst-gpios' required property set in DT */
-	if (!gpio_is_ready_dt(&cfg->reset)) {
-		LOG_ERR("Reset GPIO device %s is not ready", cfg->reset.port->name);
-		return -ENODEV;
-	}
-
-	ret = gpio_pin_configure_dt(&cfg->reset, GPIO_OUTPUT_INACTIVE);
-	if (ret < 0) {
-		LOG_ERR("Failed to configure reset GPIO, %d", ret);
+		LOG_ERR("SPI communication is not working, %d", ret);
 		return ret;
 	}
 
@@ -432,31 +258,15 @@ static int lan865x_init(const struct device *dev)
 		return ret;
 	}
 
-	return lan865x_gpio_reset(dev);
+	return lan865x_default_config(dev);
 }
 
 static int lan865x_port_send(const struct device *dev, struct net_pkt *pkt)
 {
 	struct lan865x_data *ctx = dev->data;
 	struct oa_tc6 *tc6 = ctx->tc6;
-	int ret;
 
-	k_sem_take(&ctx->tx_rx_sem, K_FOREVER);
-	ret = oa_tc6_send_chunks(tc6, pkt);
-
-	/* Check if rca > 0 during half-duplex TX transmission */
-	if (tc6->rca > 0) {
-		k_sem_give(&ctx->int_sem);
-	}
-
-	k_sem_give(&ctx->tx_rx_sem);
-	if (ret < 0) {
-		LOG_ERR("TX transmission error, %d", ret);
-		eth_stats_update_errors_tx(net_pkt_iface(pkt));
-		return ret;
-	}
-
-	return 0;
+	return oa_tc6_send_chunks(tc6, pkt);
 }
 
 const struct device *lan865x_get_phy(const struct device *dev, struct net_if *iface __unused)
@@ -482,12 +292,19 @@ static const struct ethernet_api lan865x_api_func = {
 		.phy = DEVICE_DT_GET(DT_INST_PHANDLE(inst, phy_handle)),                           \
 		.mac_cfg = NET_ETH_MAC_DT_INST_CONFIG_INIT(inst),                                  \
 	};                                                                                         \
+	BUILD_ASSERT(DT_INST_PROP(inst, spi_max_frequency) <= LAN865X_SPI_MAX_FREQUENCY,           \
+		     "SPI frequency exceeds supported maximum");                                   \
                                                                                                    \
 	struct oa_tc6 oa_tc6_##inst = {                                                            \
-		.cps = 64, .protected = 0, .spi = &lan865x_config_##inst.spi};                     \
+		.cps = 64,                                                                         \
+		.protected = 0,                                                                    \
+		.spi = &lan865x_config_##inst.spi,                                                 \
+		.timeout = CONFIG_OA_TC6_TIMEOUT,                                                  \
+		.interrupt = &lan865x_config_##inst.interrupt,                                     \
+		.reset = &lan865x_config_##inst.reset,                                             \
+		.tx_rx_sem = Z_SEM_INITIALIZER((oa_tc6_##inst).tx_rx_sem, 1, 1),                   \
+		.int_sem = Z_SEM_INITIALIZER((oa_tc6_##inst).int_sem, 0, 1)};                      \
 	static struct lan865x_data lan865x_data_##inst = {                                         \
-		.tx_rx_sem = Z_SEM_INITIALIZER((lan865x_data_##inst).tx_rx_sem, 1, 1),             \
-		.int_sem = Z_SEM_INITIALIZER((lan865x_data_##inst).int_sem, 0, 1),                 \
 		.tc6 = &oa_tc6_##inst};                                                            \
                                                                                                    \
 	ETH_NET_DEVICE_DT_INST_DEFINE(inst, lan865x_init, NULL, &lan865x_data_##inst,              \