--- a/drivers/ethernet/eth_lan865x.c
+++ b/drivers/ethernet/eth_lan865x.c
@@ -99,6 +99,13 @@ static void lan865x_iface_init(struct net_if *iface)
 		return;
 	}
 
+	/*
+	 * Interrupt line will remain as low, after completing the reset operation. On reception of
+	 * the first data header interrupt line will become high
+	 */
+	tc6->int_flag = true;
+	k_sem_give(&tc6->spi_sem);
+
 	net_if_set_link_addr(iface, ctx->mac_address, sizeof(ctx->mac_address), NET_LINK_ETHERNET);
 
 	tc6->iface = iface;