This pull request introduces support for full-duplex SPI operation in the Microchip LAN865x MAC-PHY device.

1. Separate OA TC6 specific code implementation from lan865x driver
2. Implement SPI full-duplex operation in transmit path of the LAN865x MAC-PHY
3. Change SPI TX and RX buffer size as user configurable
4. Implement SPI full-duplex operation in receive path of the LAN865x MAC-PHY