@lmajewski 2025-05-18
@RamyaI73943 Regarding the memory footprint: https://github.com/zephyrproject-rtos/zephyr/pull/89639#issuecomment-2885743314

Just for the record, because you have been using a SoC with relatively large memory (128KiB+).
On: https://docs.zephyrproject.org/latest/boards/st/nucleo_g431rb/doc/index.html

I do have 22KiB of SRAM total.

With the **Full duplex** (OA_TC6_TX_RX_BUFFER_SIZE - 68, , OA_TC6_ETH_BUFFER_SIZE - 1524) it just increases by 256B, which is ~1.2% - and this would be OK.

With the **Full duplex** (OA_TC6_TX_RX_BUFFER_SIZE - 2108, OA_TC6_ETH_BUFFER_SIZE - 1524) the RAM usage increases by: 4288 B, which is ~19% or RAM. And this is not acceptable, as the user application would be short of RAM.
That is why I've been so strict about using the large buffers.
But with the Kconfig option - it looks like it is now fixed.
(I will test this patch set in the next few days).