@lmajewski 2025-05-21
@RamyaI73943 - I've tested your patches on top of: SHA1: 53e751dcf9bbd295465c1cfe5f0b4e735d2684aa

I've strip down all the user features and set `OA_TC6_TX_RX_BUFFER_SIZE(=2108)`
When I do call:
`west build -p always -b <board> samples/net/zperf/ -- -DCONF_FILE=boards/<board>.conf`

and on the target
`uart:~$ zperf udp download 5001`

on RPI4 (with the USB <-> T1S dongle):
`iperf -l 1460 -u -c 192.168.3.3 -i 1 -b 10M -p 5001`

`[00:03:27.268,000] <err> net_pkt: Data buffer (1514) allocation failed (oa_tc6_allocate_rx_pkt:301)`
`[00:03:27.268,000] <err> oa_tc6: Device error: -12`
`[00:03:27.268,000] <err> oa_tc6: Non recoverable error`

This error I also see when I try to run [stress tests](https://gitlab.com/network851614/udp-perf/-/commits/v.1.2.0?ref_type=tags) on the board.

You can add and enable it on your side with:
`+CONFIG_UDPPERF=y`
`+CONFIG_CPP=y`
`+CONFIG_STD_CPP17=y`
`+CONFIG_REQUIRES_FULL_LIBCPP=y`

west.yml:
`+    - name: udp-perf`
`+      path: modules/udp-perf`

On the RPI4 (with dongle) you shall build the RUST program and then run:
`./target/armv7-unknown-linux-gnueabihf/release/udp-perf client --ip 192.168.3.3 --packets 100 --mode reduced --runs 1`

It will fail without finishing the first run with:
`[00:00:24.420,000] <err> oa_tc6: Device error: -105`
`[00:00:24.420,000] <err> oa_tc6: Non recoverable error`

(with the old driver it was running with it for weeks + echo_client + several other devices connected to single bus).

I've also noticed that on the STM32 the zperf shows also ~3.3 Mbps even with your patches. I cannot explain for now why it is like that (spi driver's HAL?). This needs to be investigated as well.