@RamyaI73943 2025-05-23
> I've tested your patches on top of: SHA1: [53e751d](https://github.com/zephyrproject-rtos/zephyr/commit/53e751dcf9bbd295465c1cfe5f0b4e735d2684aa)
> 
> I've strip down all the user features and set `OA_TC6_TX_RX_BUFFER_SIZE(=2108)` When I do call: `west build -p always -b <board> samples/net/zperf/ -- -DCONF_FILE=boards/<board>.conf`
> 
> and on the target `uart:~$ zperf udp download 5001`
> 
> on RPI4 (with the USB <-> T1S dongle): `iperf -l 1460 -u -c 192.168.3.3 -i 1 -b 10M -p 5001`
> 
> `[00:03:27.268,000] <err> net_pkt: Data buffer (1514) allocation failed (oa_tc6_allocate_rx_pkt:301)` `[00:03:27.268,000] <err> oa_tc6: Device error: -12` `[00:03:27.268,000] <err> oa_tc6: Non recoverable error`


@lmajewski From the error message Device error: -12, looks like there is no enough memory. (Error code 12 - ENOMEM)
Could you reduce the OA_TC6_TX_RX_BUFFER_SIZE value to smaller value such as 68, 136 or 204 and do the test again?
By the way we can confirm that whether the issue is with buffer space or something else.


> I've also noticed that on the STM32 the zperf shows also ~3.3 Mbps even with your patches. I cannot explain for now why it is like that (spi driver's HAL?). This needs to be investigated as well.

Can you tell what is the buffer size you configured here? 

From my side I did test on STM32 Nucleo F413ZH platform too. To achieve maximum throughput consider the below things.

1. Enable SPI DMA support in STM32F413ZH Nucleo-144 platform through menuconfig option. DMA support must be enabled.
```
(Top) → Device Drivers → Serial Peripheral Interface (SPI) bus drivers → STM32 MCU SPI controller driver -> STM32 MCU SPI DMA Support
```

2. In my case, in device tree spi-max-frequency is set to 24 MHz `spi-max-frequency = <24000000>;` Nucleo F413ZH's system clock is driven by PLL clock at 96MHz. Nucleo F413ZH platform's Serial Peripheral Interface(SPI) will operate in the frequency based on the bit values in (BR[2:0]: Baud rate control) SPI_CR1 (SPI Control Register 1) register. In Zephyr, Nucleo F413ZH platform's SPI driver code (/drivers/spi/spi_ll_stm32.c) sets the Baud rate control bit value based on the  spi-max-frequency value given in nucleo_f413zh.dts file. The possible frequency values for SPI that operates is,
```
000: 96/2
001: 96/4
010: 96/8
011: 96/16
100: 96/32
101: 96/64
110: 96/128
111: 96/256

```
if `spi-max-frequency = <15000000>; `is set to 15 MHz, the SPI will operate in 12MHz (96/8 => 010)
if `spi-max-frequency = <24000000>;` is set to 24 MHz, the SPI will operate in 24MHz (96/4 => 001)
So, After 12 MHz, available maximum frequency value is 24 MHz.
At 12 MHz achieved bandwidth is 8.78 Mbps. At 24 MHz achieved bandwidth is 9.46 Mbps.
So it is better to set the higher frequency value in your device tree file to achieve maximum throughput.