@lmajewski 2025-05-24
@RamyaI73943 
- I've reduced the `OA_TC6_TX_RX_BUFFER_SIZE=68` and I do have the same behavior.
`<err> oa_tc6: Device error: -105`
`<err> oa_tc6: Non recoverable error`

Now the error is different `-ENOBUFS`.

I've spend some time and connected the LAN865x to STM32G491 and it even with `OA_TC6_TX_RX_BUFFER_SIZE(=2108)`
has enough resources to run this driver without issues (`echo_server` + `udp-perf`) :
`Memory region         Used Size  Region Size  %age Used`
`           FLASH:      276120 B       512 KB     52.67%`
`             RAM:       79676 B       112 KB     69.47%`

So the error shall not be present.

I also had to increase the `CONFIG_OA_TC6_IRQ_THREAD_STACK_SIZE` to 1024B for having echo_server running,
Increasing it further to 2048 bytes did not helped with the described above issue.

Your code has been tested with (the `udp-perf` is supported as a module in both below cases):
`west build -p always -b <board> samples/net/zperf/` and
`west build -p always -b <board> samples/net/sockets/echo_server/`

There are issues with memory/buffers allocation even with `OA_TC6_TX_RX_BUFFER_SIZE=68`

I think that, this issue shall be fixed first with the "basic" SPI operation (i.e. DMA is NOT enabled).

Can you share your test procedure, so I can run it on my setup?
Can you use the `udp-perf` on your side?