@RamyaI73943 2025-05-27
Hi @lmajewski, I am trying to reproduce the issue you mentioned above. For that I am using Nucleo-L432KC which is available on my desk that has less memory resource. (64 KB SRAM and 256 KB FLASH).  It looks to build Zperf application in these platforms, we have to make some configuration changes in menuconfig (To reduce the memory space used). 

Can you share the details about what menuconfig configurations are you using while building the Zperf application. If I am not wrong, you are using .conf file which override/sets the required configurations to build and run the Zperf application. Can you share that file?