@lmajewski 2025-05-27
@RamyaI73943 - 64 KiB of RAM is on the limit.
My build with zperf uses:
`58684 B       112 KB     51.17%`

I've just send the config file via email.

Moreover, I just would like to emphasis, that setting `upd-perf` would probably show you errors almost immediately.