@lmajewski 2025-12-04
@RamyaI73943 
Please find some performance report regarding this patch set:
Test setup:
###########
Nucleo STM32G491RE
`west build -p always -b nucleo samples/net/zperf/ -- -DCONF_FILE=boards/nucleo.conf`
OA_TC6_TX_RX_BUFFER_SIZE = 31 (Full duplex patches)

1. Zephyr module (nucleo board) as CLIENT

`zperf udp upload 192.168.3.1 5001 10 1460 10M	`		-> nucleo shell
`iperf -s -l 1460 -u -B 192.168.3.1 -i 1` 			-> RPI4

2. Zephyr module (nucleo board) as SERVER

`zperf udp download 5001`
`iperf -l 1460 -u -c 192.168.3.3 -i 1 -b 10M -p 5001`


With full duplex:
#################
Ad 1.
Rate:                   3.04 Mbps       (3.16 Mbps)
Ad 2.
rate:                  3.36 Mbps

Current driver:
###############
Ad 1.
Rate:                   2.91 Mbps       (3.03 Mbps)
Ad 2.
rate:                  3.25 Mbps

There is some small performance improvement.

As far as I remember you are using the 'SAM5' board? And your results were around ~10 Mbps?
Can you confirm this?

Apparently HAL for SAM and STM32 differ - especially terms of the SPI performance ... (as partially pointed out in https://github.com/zephyrproject-rtos/zephyr/issues/74393)

Tests for stability are ongoing...